+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added [17/10/23]
+- shuffleTiles method to player.py
+- score attribute in player class initialization
+- score keeping in each player instance in putWord method
+- test_putWordGetScore in test_scrabble_game.py
 ### Added [16/10/23]
 - fillTiles function in player.py
 - tests in test_player.py